Testing the build. Added main.c to the makefile.
I also fiddled around a bit with the optimisation level setting in the cflags.
Interestingly, setting it to lower optimisation results in a dramatically
smaller hex image. I believe the higher optimisation level results in some
function calls being removed (and function code being put in place), which
results in larger compiled code if the function is called from a few different
places.