Part way through debugging.
The problem is that no message comes through on the serial console (I've
tried with screen, picocom, and even the Arduino serial console). I've played
around with the position of the logic to switch off the status LED, by which
I've determined that the interrupt logic seems to be working correctly.

I'm thoroughly confused why no data is coming through to the USB console.