Moved the push button signal to a different pin.
This keeps the same hardware arrangement as for the Arduino DigitalReadSerial
example.

I've done some further debugging by loading the actual Arduino example, and I found
that it works, and data is displayed in the Arduino serial console and also
through the screen command. This means that the hardware is set up correctly,
and I also now know that I'm invoking the screen command correctly. There must
be a problem somewhere in the MCU code.