Finally got the serial to work!
From knowing that the hardware setup worked with Arduino, I altered my code
to bypass the interrupt driven USART logic, and when I still wasn't getting
data on the tty, I realised it was something in my USART initialisation code
that I was doing wrong, but the Arduino code was doing right.

I had to delve into the Arduino Serial source code, where I discovered that it
was doing some calculation with the baud rate to get a value to assign to the
baud rate register. Rather than assigning 9600 to UBRR0H:UBRR0L I actually had
to assign 103. And hey presto, it works.

Incredibly confusing, because the calculation logic that I found in the Arduino
code doesn't seem to be mentioned in the ATmega328P datasheet, even though it's
clearly required.