Changed the way messages are buffered by the USART driver code.
Previously it copied messages into a fixed length char array (32 chars long)
but I noticed that sometimes the messages in the serial console would be a
bit mangled if there were a few messages sent very close together (due to
switch contact bounce). My thinking is that the mangling was caused because
the buffer was filled, and one message was only partly copied.

So I've gone from a fixed length array of chars, to a fixed length array of
pointers to strings. Messages should now be either sent or not sent (if the
buffer has no space for another pointer to string, which should be unlikely).
The logic for adding a message to the buffer had to be changed, and also the
UDRE interrupt handler had to change, so that it now effectively steps along
the string pointed to from the buffer until it reaches a null byte.