Completed work on the analog_read function.
I found that I had to add an ISR for the ADC interrupt, as it's mentioned in
the datasheet (at least, it's necessary in order to use the ADC noise
reduction sleep mode).

Also added the C sources in the makefile.

I test compiled analog.c, but I've realised that I'll have to add some more
code in uart.c in order to convert numbers to strings. I've checked the AVR
GCC man pages which confirms that printf definitely is *not* available, lol.