Major redesign of uart.c to handle output of different data types.
Most significant is the change to the data structure used to buffer messages.
This is no longer a simple array of string pointers. It's now an array of
structs, each containing a union (which maps to the data type being output,
either char* or int values), and a pointer to a function to print a single
character of the data, and advance to the next character (returning 1 if
all finished).

As of this commit, I've converted the transmit_string function to work with
this new buffer structure, and I've also updated the ISR code so that it
invokes the output function pointer, and updates the queue pointers according
to the return value.

I still have to implement the functions to perform int output.