Got the code to correctly transmit one value per timer IRQ.
I still haven't figured out why the ADC ISR is continuously triggered. I've
even tried writing the interrupt flag, but that didn't change it. It produces
garbage because it fills up the USART message buffer with a flood of messages
faster than the USART hardware can transmit them. Once the buffer is full,
further messages are dropped.

I'm actually not entirely sure that the current code is getting new values from
the ADC, or if it's just printing the same value over and over. I will have to
set up a circuit with a thermistor to test this.