Started work on the Fade demo.
Added the makefile (same as from the AnalogReadSerial demo) and also created
pwm.h which declares some functions to use for creating a pwm signal.

For the first version of this, I'm only going to use the OCR0A and OCR0B pins,
because they will be set automatically by the timer hardware. I think it
should also be possible to generate PWM signals on any pin by making use of
the output compare match interrupt vector.