Did an experiment to compare Arduino analog performance with my implementation.
The results are surprising. My implementation, which uses the noise reduction
sleep mode, produced 1,000 readings, all of them 512. I verified that new
readings were being taken by disconnecting the A0 pin and connecting it to
ground and 5V, and the analog reading changed accordingly.

When I compiled the Arduino sketch and collected 1,000 readings from that
however, two things stood out: there was more variance in the analog readings,
and also, the mean was not close to 512 as I would expect given a voltage
divider circuit provided 2.5V to the A0 pin. Instead the mean was around 519.

It appears that the noise reduction sleep mode significantly reduces the
variance in readings collected from analog inputs.