Added main.c and implemented the main function.
In testing there were a couple of very challenging bugs. I didn't realise at
first that the PWM output pin must be set to output mode, so I was very
confused for a long time why the LED wouldn't light up.

There was also a logic error in shifting the compare output mode bits for
channel A or B.

According to the datasheet, generating the PWM waveforms via the compare
interrupt isn't a good idea because it would generally take too much CPU time,
leading to corrupted waveforms.