Added a function to tell how many spaces are free in the UART tx queue.
This is very useful to allow the caller to avoid trying to send a batch of
messages if there isn't enough space in the queue for all of them. Previously,
when the queue was filled due to messages generated faster than the UART
hardware can transmit, messages would be dropped, and the result was garbled
text in the serial console (eg, all the newlines are dropped).