Tested the program on the ATmegaM328P chip.
One problem that I found was that the serial console seems to treat the enter
key as a \r character, without a line feed \n. As a result, I had to modify my
getline function to look for \r as the line ending.

Also, when printing the string, I had to add another statement to print a \n
character, otherwise the cursor doesn't advance to the next line, and the next
printed line over-writes the previous line.