Finished debugging the I2C demo.
The main problem was that I had to set the bus speed (TWBR register).
There are a few other changes I made, particularly in how I set the bits
in the TWCR register (I found out that there's an easy way to get the
bit mask from the AVR header files).