After much debugging, got a basic version to work.
This version doesn't make use of the interrupt capability of the MCP23008
chip; instead it continually polls the state of the MCP23008 GPIO register
over the I2C bus.

Most of the trouble I had in debugging this was due to an obscure problem
relating to how the MCP23008 handles REPEAT START signals on I2C. Basically,
if you want to write to several registers that aren't adjacent to one another
you must send STOP, then START, then chip address, then register address.
If you send a REPEAT START, the MCP23008 assumes you will write to the next
register; and so it was writing what I sent as a register address into one
of it's registers.

I fixed this by changing the dequeue function to tell the AVR hardware to
send a STOP then START instead of a REPEAT START, which has solved the
problem.