Moved the template makefile to Makefile.example.
I've made changes to the makefile in the library directory to add the
capability of compiling the source files in the library dir into a statically
linked library, which I've called libavrutils.a.

I've also added a rudimentary install target to the makefile which will
install the header files and libavrutils.a to the system include and lib
directories (by default, under the /usr/local prefix). I believe this should
enable the headers and lib to be found by avr-gcc without having to provide
any additional search paths via the -I flag.