Updated the template makefile to handle libs.
Unfortunately, the avr-gcc toolchain doesn't seem to have /usr/local/include
and /usr/local/lib on the default search path, so I've added them in the
template.

I've also modified the template so that when the target ELF image is linked,
any static linked libraries come after the object files, as is required.