Added code for managing a queue of data to the SPI
This consists of the logic to add data to the queue, take data off the queue,
and maintain a list of unallocated queue slots. Also code to initialise all
the queue slots in the unallocated list.

Added logic in the SPI interrupt handler to correctly handle taking data off
the queue and placing it in the data register.