Working version of the simple test.
In order to get this to work, I had to scale back my ambition for interrupt
driven SPI logic. Instead, to get working code off the ground, it now polls
the flag in the SPI status register to find when the transfer is done.

I've removed the queue maintenance logic (not sure why it wasn't working as
intended).

I've also changed the code to initialise the LCD panel, based on code that
Adafruit implemented in their Arduino library.