Demo to draw small rectangles.
I've also made some further changes to lcd.c, to allow us to set a rectangular
display window and fill it with a colour.

Most bugs are now sorted out, but lcd.c needs to be tidied up a bit. It would
also be good to separate the hardware dependant code from the generic graphics
code, as that would make it simpler to re-use code with a different display
controller chip. Needs to be balanced though; the DFRobot libraries are great
for supporting different hardware, but they take up lots of flash to the point
that the full set of libraries don't fit on the M328P's 32KB flash. And there
are plenty of AVR chips with tighter flash constraints.