Quick demo to work with a rotary encoder.
Very simple, just print the signal states over the uart. The hardware seems
to have a bit of an issue with switch bounce inside the rotary encoder itself,
so I might consider how to resolve that. Either a delay, or I'm wondering
about using a capacitor between the encoder signal and ground.