Print the rotation direction over the uart.
I've changed it so that the interrupt is only triggered by channel A, since it
seems that both lines change once per detent. Added logic to find the
rotation direction from the A/B states, which is quite simple.

Also added debouncing, by having a 5ms delay before we get the pin states. This
seems to work well enough most of the time.