Linked list for UART buffer.
The transmit queue list has two pointers, to the head and tail of the list.
That allows us to efficiently remove items from the head, and also add items
to the tail.

There is also a linked list of unused buffer slots; only need a pointer to the
head of that. We add and remove items to the head of the free list, since every
unallocated slot is the same.

Compiles successfully, now on to testing.